Created WildcardNode and made adjustments to ScriptNode
Removed ScriptNodeUI beucase of the impossibility of the use of generics. To replace it, WildcardNode(and UI) was created(may or not be a temporary name). It serves as a placeholder for a node and can be replaced by any kind of node. It is useful mainly with ScriptNode given that there is no ScriptNodeUI. AnimationGraphNode now requires that Script Nodes are added through an specific method called AddScriptNode due to the necessity of using generics. Added events to AnimationGraphNode when nodes inside it are modified in some kind, like being added, removed, connected and etc.