Changed State Machine Connections. Created NamedItemGroup and an UI Finder. And More
Almost all changes were made on the State Machine and it's UI. Turned most of TransitionConnectionUI into ConnectionUI, implementing transition specific stuff in TransitionConnectionUI which inherits from ConnectionUI. Created NamedItemGroup to use when a group of objects need to be uniquely named and implemented it in the state machine UI and partially on the state machine node. Created a Finder UI that works with NamedItemGroup, allowing to select items contained in it. Removed ports and edge totally from StateMachineEditor, switching to the Finder UI. Changed how lists inside nodes are implemented, such as transition list and any state filter list.